- refactoring commands, moving all commands into command.rs
